Fix bug on FileServer converting QString to const char* (#10187)
`toUtf8()` returns an object which will be destroyed as soon a the call finishes. By assigning its data to a `const char*` it keeps a reference to a deleted object. Fixed by passing it as parameter to the function, that way the data will be valid until the function finishes.

Signed-off-by: moraaar <moraaar@amazon.com>