Changed one of Area Light bus's Set Intensity Methods to a Different Name
Changed to SetIntensityAndMode. O3DE's EBus system does not work with methods that have multiple overloads, so either we would need to change the EBus system to work with multiple overloads or change the Area Light Bus to not have multiple overloads of a single method. I went with the latter, as it's an infinitely easier change.

I also took the liberty to add this as an extra method with tooltips for Script Canvas, as it did not exist as a proper script canvas node yet.

Signed-off-by: Abis <26350896+TheOneAbis@users.noreply.github.com>