Unify fast math across all platforms (#19303)
* Fixes inconsistent fast math settings across all platforms.

* Fast math is off by default, but is supported to switch on
* Fast math compile options that are platform independent
  have been added when fast math is switched on.
* Extra CMake variables have been added which allow individual target
  to turn off fast math if necessary (for 3rd Party)

This is commit 1 of 2 - this first commit is all the cmake changes.
Commit 2 will be the changes to code to use the new compile options,
as well as fix any warnings.

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

* Fast math support - part 2 - fixes to cpp and h code

Turning fast math off enables detection of type casting
values from double to float accidentally, and will issue
a precision warning.

This fixes those warnings (by fixing the actual code, not by
disabling the warning)

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>

---------

Signed-off-by: Nicholas Lawson <70027408+nick-l-o3de@users.noreply.github.com>